admin: remove `actionObject` usage in APIs
Currently in the admin API, the methods that deal with adding or
modifying an action are passed an `actionObject` which contains multiple
fields including the target of the action, the event type of the action,
and the `actionToSet`.

This behavior is inconsistent with the rest of the API methods, which
always take separate parameters and never several ones contained in a
single object.

Considering that the use of separate parameters makes the API clearer
and serves as self-documentation, this commit standardizes all methods
by removing the `actionObject` and instead passing the various separate
parameter.