admin: rewrite `call` using exceptions instead of error values
The current `call` method, which is used to call the server's APIs,
returns an array of two values where the first value is the potential
value returned by the server and the second value is the potential
error.

This return value allows for destructuring in the caller (`[value, err]
= await call()`) and for checking the error value before proceeding with
execution (`err != null`).

This syntax was taken from Go and is not a syntax typically used in
Javascript. Moreover, during the implementation of type checking with
Typescript, this syntax causes difficulty in typing the return value of
the APIs because the value in question is not directly the value
returned by the server but rather an array that can contain both the
return value and an error.

For this reason, `call` has been rewritten with exceptions, which are
much more idiomatic and used in Javascript and allow for more precise
typing of the values returned by the APIs.