admin: improve `ConnectionOverview` route
The `ConnectionOverview` route originally contained the feature to
import the users of a connection.

Subsequently, with the introduction of actions, this feature was moved
to the `Import users` action and therefore the `Start a new import...`
button in `ConnectionOverview` stopped working.

With this commit, the aforementioned button is removed from the
`ConnectionOverview` route and the layout and design of the route is
slightly retouched to make it more in line with the rest of the
application.

Also, the grid showing the list of imports is now rendered using the
`Grid` component, instead of using a basic implementation of a grid.