test: use a temporary directory for the storage
Currently, the test `TestExportUsersToFile` uses a directory within the
repository for creating temporary files. This is uncomfortable when
working with Git, that may, in case of bad configuration of `.gitignore`
files, inadvertently track the temporary files.

So, this commit instantiates a storage within the system's temporary
directory, which is removed when the test finishes successfully.

Note that the directory `test/testdata/storage` is no longer used for
creating temporary files, so the `.gitignore` files referring to it can
be cleaned.