apis: fix duplicate running scheduler
Currently, after a leader election, if the same leader is re-elected, a
new scheduler is created without closing the previous one.

With this change, in such a situation, a new scheduler is not created,
and the previous scheduler continues running.

Additionally, this change ensures that listeners on the `ElectLeader`
notification are invoked only when the leader has actually changed.