apis/postgres: fix and improve the `ListenToNotifications` method
Previously, the `ListenToNotifications` method incorrectly assumed that
functions would consistently return the `context.Canceled` error when a
context is cancelled. However, functions can potentially return various
errors in this scenario. This change fixes this issue by checking the
`Done` channel to determine if a function has returned an error due to
context cancellation.

Additionally, this change refactors the method to return the
notification channel and the associated stop function directly, instead
of returning them within a struct.