test/chichitester: improve checking for the cancel of Chichi
Currently, the check for the cancel of Chichi is done by checking the
error returned by `launchChichi`. This is inaccurate and does not work
on Windows (see the issue #233 for more details).

This commit improves the checking for the cancel of Chichi by checking
the `ctxWithCancel` context itself.

It also removes the function `isSignalKilledError`, which becomes
unused.

Fixes #233.