admin: fix the generation of the action's output schema
Currently, when saving an action of a source connection, its output
schema is composed of the properties used in the mapping (if a mapping
has been defined).

Additionally, when the action uses a nested property, the output schema
should not contain the property itself but rather the top-level object
from which the property descends.

However, in the current action-saving algorithm, this principle is not
met, and nested properties are also included in the output schema,
subsequently causing server-side validation errors.

This commit fix the issue.