trace-events-script: always check Prettier version before formatting
The `trace-events-script` directory includes a web tracing script and a
mock website for testing it. Currently, we are using Prettier to ensure
all the files in the directory maintain a uniform format.

The formatting is done by using the two terminal commands `prettier` and
`prettier-test-website`, listed in the `package.json` file of the
directory.

With this commit, before you run these commands, the system will check
if there's a newer version of Prettier available. If so, you'll get the
option to download it before starting the formatting.