apis/postgres: fix compressed notifications
If a notification is compressed due to its length, but it can be
successfully sent with a single `NOTIFY` call, the recipient may fail to
decode it because they expect it to be uncompressed.

Additionally, this commit fixes a bug that compressed the payload before
removing the newline added by `json.Encode` and before escaping it.