apis: update the `ReloadSchemas` method to reload all schemas
This commit modifies the `(*Workspace).ReloadSchemas` method to reload
schemas for users, groups, and events. If a schema cannot be reloaded
due to a missing table, it will not reload any schemas and will return
an unprocessable error with the code `TableNotFound.`

Note that even though the events schema is reloaded, it is currently not
used.

Additionally, this commit changes the return type of the
`(*Store).Schemas` method to return a map instead of individual schemas.