apis/datastore/warehouses: properly close rows of the Query method
Note that in the "github.com/jackc/pgx/v5" module, it is necessary to
close the rows before calling the `Err` method. However, in the
"github.com/ClickHouse/clickhouse-go/v2" module, the documentation does
not specify when to call `Close`, but it appears to be acceptable to
call it before the `Err` method.