commit: run `npm install` when executing `go run commit/commit.go`
Currently, the command `npm install` must be manually executed in the
admin. This means that, for example:

* when launching Chichi, nothing ensures that the local versions of
  packages are up-to-date, and this may lead to unexpected behaviors.
* when committing to the repository, the file `admin/package-lock.json`
  is not checked to be up-to-date.

So, according to the logic we already use with `go.mod` and `go.sum`
files, this commit ensures that the package versions are synched with
the versions indicated in the file `package.json`, increasing
consistency of behaviors and reproducibility.