server: fix the endpoint that exposes `AddWorkspace`
Commit b949b95 introduced the workspace ID in the HTTP API endpoints.

Erroneously, the endpoint exposing the `AddWorkspace` method was also
modified to include the workspace ID, thus becoming a POST on
`/api/workspaces/{id}`. However, the `AddWorkspace` method does not
require any workspace ID to be executed.

This commit resolves the issue by restoring the endpoint, which is now
once again a POST on `/api/workspaces`.