server: improve the handling of the workspace related HTTP APIs
This commit addresses issues related to the logic of HTTP APIs
concerning workspaces.

Previously, the logic did not correctly handle endpoint paths that did
not have a trailing slash, specifically those of the type
"/api/workspaces".

Additionally, prior to this commit, the logic always expected the
workspace ID to be present in the endpoint path. Now, the logic is
equipped to correctly process both basic endpoint paths
('/api/workspaces' and '/api/workspaces/') and those that indeed include
a workspace IDs.