all: implement workspace switch
This commit introduces the workspace switch in the UI. Now, regardless of the
page the user is on, they will be able to switch the workspace to view the data
associated with it.

Upon login, a list of workspaces will be displayed so that the user can select
the workspace they want to view (unless there is only one workspace, in which
case that one will be used directly). On this screen, it is also possible to
create new workspaces.

In the settings page, a "General" section has been added where the user can
modify workspace information (name and privacy region) as well as delete the
workspace.

Since it is now possible to create new workspaces but they are not automatically
associated with a data warehouse, indicators and messages have been added to the
UI prompting the user to connect the data warehouse before proceeding with
certain operations.

Finally, hardcoded logic that previously always used the workspace with id 1 has
been removed.