admin: re-instantiate Monaco editor after language change
Currently, in the action screen, after creating a JavaScript
transformation function and switching to Python, a syntax error is
displayed by the editor (which doesn't happen the other way around when
starting with Python and then switching to JavaScript).

This error seems to be related to the fact that when using JavaScript,
Monaco editor internally keeps additional information that is not simply
erased by changing the language.

This commit therefore forces a new instance of Monaco Editor every time
the language is changed, solving the problem.