apis: validate the action filter
This change validates the filter of an action when an action is added or
changed. The `(*Connection).AddAction` and `(*Action).Set` methods
return a bad request error if the filter is not valid.

It also fixes a bad request error that occurs if a property is used in a
filter but not in the mapping.