apis: add write keys to mobile and website connections
This commit extends the functionality of write keys to include mobile
and website connections, in addition to server connections. It also
refines how events are sent from the analytics SDK by using the write
key as connection identifier.

When collecting an event, the system now verifies the presence of the
write key in both the Authorization header and the request body's
writeKey field.

Additionally, this commit changes event listeners by removing
server and stream connections.