apis/mappings: validate required transformed properties
The `Apply` method must verify that a transformation returns all
required properties. For actions using a mapping instead of a
transformation, a subsequent commit will validate the existence of a
mapping for any required property when an action is created or updated.

Additionally, this change refactors the code that validates the values
returned by a transformation.