apis: support nil `Priority` and `Mapping` when setting anonymous idents
Currently, when calling the `SetAnonymousIdentifiers` method with nil
fields for the `ids` structure, the method returns an internal error
indicating that a "not null" constraint on the database has been
violated (this happens for `Priority`).

This commit modifies the method to treat the fields of `ids` the same
way whether they are nil or empty.