apis: changes the way ext. identifiers for files and dbs are handled
In particular, the property serving as the external identifier (the
first one for files, the one called "id" for databases) must have a type
of Text, JSON, or an integer type.

Note: the JSON type is supported because, when importing from a JSON
file, all properties have a JSON type. Otherwise, it would currently be
impossible to import from a JSON connector.

The conversion to a string currently takes place by calling the
`fmt.Sprint` method and passing it the value returned by the database or
the file.

When we consider optimizing the code that handles identities, we may
also review this part and any transformations in the type of external
identifiers.