apis/transformers: add exception name to the error returned by Python
At the moment, when an exception occurs in Python, only the exception
message is captured, and the type is lost. This is problematic for
exceptions of type, for example, `KeyError`, because the exception
message only contains the name of the key not found, making it
incomprehensible.

This commit adds the exception type to the error message.