apis: check that actions do not have both mappings and transformation
As long as the identifiers were at the action level, it was correct for
an action to have both the transformation and the mappings, where the
mappings referred to the identifiers.

Now that the identifiers are at the workspace level, mappings are
excluded with the transformations. Therefore, this check must be
performed when creating or setting an action.