admin: fix import from database
Currently, when a query is modified in an import action of a database
source, the mappings are disabled and an error message is displayed,
asking to confirm the new query before proceeding.

However, upon confirming the new query, the error message is not removed
and the mappings continue to be disabled.

This commit fixes the issue.

Closes #297