apis: allow empty mappings in `validateActionToSet`
Since identifiers are now defined at workspace level, there may be
actions with no properties mapped.

The discussion on whether this can be useful or not is in issue #311;
this commit currently removes the check because there is no reason to
limit actions in this way, as identities are still perfectly valid if
constructed without any properties explicitly mapped by the user.

This makes the implementation more consistent and simplifies the
development of Chichi; if in the #311 discussion we decide to limit the
creation of actions, then we will reintroduce this check in a deliberate
and consistent manner.