connectors/excel: allow reading property names from an Excel header
Previously, property names were 'column1,' 'column2,' etc. With this
commit, the names are taken from the first row of the Excel file or
generated in Excel format: 'A,' 'B,' etc.

It also renames the `columnIndexToPropertyName` function in the CSV
connector to `columnNumberToName`, as it is named in the Excel
connector.