admin: automatically update user table when a new column is added
Currently, if the schema of the user warehouse is modified, the user
table displayed in the UI is not updated, because the UI is based on the
columns previously retrieved and enriched with additional information
related to the user experience (for example, if a column has been hidden
in the table view).

This commit ensures that the updated schema is always requested via API,
every time the user table is displayed, merging with the columns saved
in local storage (if present) to add any user-set information.

In this way, the user table is always synchronized with the database
schema and potential bugs are also avoided, should a column be added
that has a special purpose for the UI (for example, the column with the
identifier also used for navigation purposes to view the user traits
page).

Closes #181, closes #319