admin,apis,database: make stats refer to "user identities", not "users"
The user import statistics have remained unchanged for some time.
Since then, identity resolution has changed, and therefore the statistic
indicating "how many users have been imported" is incorrect. This commit
changes the meaning of the statistic from "users" to "user identities",
aligning it with its usage, and modifies the UI to display different
text to refer to the value.

However, the rest of the UI has not been modified (e.g., TypeScript
variable names), and the meaning and usage of statistics as a whole
still need to be reviewed.

The discussion is taking place in issue #293.