admin: fix transformation function being overwritten in the action
Currently, when the action screen is opened, the transformation function
is always overwritten with the default transformation function.

This commit solves the problem by adding an additional check, which
ensures that the default transformation function is used only when the
programming language in use is changed or when creating the action and
switching to the transformation function mode from the mapping mode.