admin: remove old check that prevents actions to be saved
Currently, on the action screen, there are some checks that disable the
"Save" button to prevent saving the action in case there are specific
errors.

However, one of these checks is now obsolete and does not take into
account the fact that some actions might not satisfy it, even though
they are valid. This causes, in some cases, the inability to create a
formally correct action.

This commit removes that condition.

Closes #296