admin: fetch action sheets on mount
Currently, when creating actions on Excel, you must specify the path and
then the sheet, which can be selected from the list of sheets available
in the file, read via API.

However, when opening an already created action, even if it internally
has a set sheet, the corresponding field is not filled out, as the
reading of the available sheets in the file has not yet been done, and
therefore the sheet select has no selectable options.

This commit performs the reading of the sheets of already created
actions during the opening of the editing screen, so that the
corresponding field is already filled out with the correct value.

Closes #342