apis/datastore/warehouses/postgresql: simplify the matching function
Remove the possibility of returning null values from the
`matching_func`. These were necessary before making the modification
that placed the identifiers at the workspace level because two users
could be considered equal for one action, but another action could claim
not to know.

Now, this no longer occurs: either two users are the same, or they are
different. The case of "not knowing", previously expressed with null, is
now simply considered "different".