connector: pass only first level properties to `Users` and `Groups`
This change only passes the proprieties at first level to `Users` and
`Groups` methods.

Only a select few app APIs permit the passing of properties to return,
and as of now, we are only aware that MailChimp allows passing
properties beyond the initial level.

Moreover, ChiChi currently fails to handle properties beyond the first
level correctly when they are passed to the `Users` and `Groups`
methods. Given these circumstances, this change eliminates support for
properties beyond the first level that are passed to connectors.