admin: show action types when there is no action in the actions list
Currently, in the actions list of a connection, if no action has been
created yet, the "Add action" button is shown. When clicked, it opens a
dialog with a list of action types, leading to a configuration screen
for a new action.

This commit ensures that, if there are no actions, the action types are
displayed directly on the page without the need to open a dialog.

On the other hand, if at least one action has already been created, the
"Add a new action" button displayed above the list will open the dialog
as it did before.