admin: restore scroll to mapping after confirmation of file information
Currently, when the path and sheet (if present) are confirmed in a file
action, the Javascript automatically scrolls to the mapping section.

However, if one is in the action creation phase, the mapping section is
not yet displayed because it will be displayed only after confirmation.
Therefore, an error occurs during scrolling because the DOM element
corresponding to the mapping section does not yet exist.

This commit adds a slight delay before scrolling to ensure that this
occurs only when the section has been effectively inserted into the UI.

Closes #343