admin, trace-events-script: fix execution of Prettier
The current execution of Prettier is incorrect because pattern globs
determining which files to format relies on the system shell. This
results in different behavior from one system to another.

This commit escapes the patterns so that Prettier resolves them.

Additionally, a file `trace-events-script/.prettierignore` is added to
prevent formatting files within `dist`, which should be left as they
are.