admin: fix overlapping language submenu in transformation editor
Currently, when clicking on the language dropdown in the minimized
transformation editor, its submenu opens, and at the same time the
transformation fullscreen view is opened, causing the submenu to overlap
with the page content.

This commit ensures that in such cases, the submenu of the minimized
editor is not displayed. Instead, it waits for the transformation
fullscreen view to open and then displays the submenu of the fullscreen
editor, preventing overlaps but still allowing the user to choose the
language, as requested.