apis: ensure privacy region is passed when serving UI
Previously, when serving the UI while creating a new connection, the
privacy region of the workspace was not properly passed to the app
connector. This commit addresses this issue by moving the
`(*Connections).ServeUI` method to the `*Workspaces` type and ensuring
that the privacy region is correctly passed to the app connector.