admin: restore login autocompletion on Firefox
Currently in Firefox, the login form does not autocomplete with previously
entered data.

Upon investigating the problem, we found that there is a related open issue in
Shoelace that has not yet been resolved due to lack of browser support
(https://github.com/shoelace-style/shoelace/issues/269).

This commit therefore restores the native browser inputs to enable autocomplete,
styling them to be consistent with the Shoelace style. Eventually, Shoelace
inputs can be reused should the issue be resolved.