admin: always redirect base URL navigation when user is logged in
Currently, when a user types the base URL of the admin
(https://[authority]/admin/) into the browser's address bar and is
already logged into the application, a client-side redirect to the URL
https://[authority]/admin/connections occurs to display the connections
map instead of showing the login form again.

However, this redirect only happens when the base URL is manually
entered into the browser's address bar.

In fact, if the user manages to navigate to the base URL without
reloading the browser, using for example the browser's "back" button
after logging in, no redirect is performed and the application skeleton
is displayed with the central content empty, as the routing is not
configured to handle such a URL when the user is logged in.

This commit solves the problem by performing the redirect even when
navigating to the base URL without a browser reload.