connector: don't panic if an event type does not exist
At present, the `SendEvents` and `PreviewSendEvents` methods panic if
the event type does not exist. However, there is no assurance that event
types will remain unchanged. The `EventTypes` method may return varying
event types over time, as the app's event types are not fixed but
subject to change, or a new version of the connector might remove
support for certain event types that are no longer exit.

This situation also implies that a destination action associated with
events may no longer be able to send events if its corresponding event
type no longer exist.

With this commit, the `SendEvent` and `PreviewSendEvent` methods have
been modified to return a specific error if the event type does not
exist, instead of panicking.