apis: fix destination app action creation by correctly validating schema
When creating a destination app action, the input schema of the action
to be created will be the schema of the events. Therefore, the input
schema passed to the 'AddAction' method is expected to be the invalid
schema. However, the `validateActionToSet` method does not consider this
case and marks the passed input schema as invalid.