admin: show the executions list of the destination connections
Currently, in the "overview" page of the source connections, the list of
action executions is shown.

This commit ensures that the same list is also shown in destination
connections.

Note that the title of the execution list and, in general, the term used
by the UI to refer to an execution differs based on the role of the
connection: in the case of sources, the term used is `import`, while in
the case of destinations, it is `export`.