commit: bypass the cache of go tests and execute them every time
Passing the option `-count 1` ensures that every test is executed every
time, preventing go test from considering a test "successfully" even if,
in fact, it has not been executed.

This prevents the case when a source file which is not a Go file (eg. an
SQL script) has changed, as such change cannot be noticed by go test.