admin: fix the double tooltip on the connections map's status dots
Currently, when hovering over a status dot of a connection in the
connections map, the tooltip that describes the status of the connection
is shown twice (one tooltip is positioned on top of the other).

This behavior is caused by a bug in the code, which wraps the status dot
with the same tooltip twice, in two different components.

This commit fixes the problem.