all: remove the layout from types and add app specific formats
This change is preparatory for the modification that will normalize the
data read by the apps using the app schema instead of the action schema.

It removes the layout field from types. This way, the `DateTime`,
`Date`, and `Time` types no longer contain any information about the
layout. The layout is now only maintained at the level of individual
connectors and only for apps.