admin: implement fullscreen mapping
With this commit, we consolidate the two action editing modes, namely the
transformation function and mapping, into a single editing box.

The content of the editing box is dynamic, accommodating either the source of
the transformation function or the mapping. This adaptability is determined by
the user's selection of the current editing mode through the button groups
located in the header of the box.

In addition, the header features a "Fullscreen" button. Upon clicking, it
initiates the opening of the full-screen view. In this expanded view, the two
columns on the left are designated for the execution of the transformation and
the display of property lists.