admin: add workaround to show matching properties in the action screen
Currently, when opening the editing screen of an already created action
that possesses matching properties, the fields of the matching
properties are populated with the value `undefined`.

The issue arises because the server serializes the data structure of
matching properties inconsistently compared to all other cases where it
sends properties to the UI. Specifically, the `name` field, which
usually contains the property name, is serialized as `Name` in this
case. Consequently, accessing `name` results in obtaining the value
`undefined`.

This commit introduces a temporary workaround by converting the property
type to `any` with a type assertion. This allows access to the `Name`
field and correctly displays the names of the properties without
triggering TypeScript errors.

This code should be removed once the server behavior is fixed.

Refer #381