apis/connectors: check for schema conformity before normalization
When reading users from an app, the data is normalized using the schema
provided as an argument to the Users method. However, the schemas may
differ from the app schema, potentially leading to data corruption.

With this change, we first verify whether the provided schema conforms
to the app schema. It conforms if it is essentially the same as the
app's schema, allowing it to be used for normalization.

Additionally, this change implements a caching mechanism for app schemas
to avoid reloading them for each call to the `Users` method.

For #380