apis/commits: add the `Records` method to the `Database` type
This change introduces the `(*Database).Records` method alongside the
`Query` method. The `Records` method accepts a schema as a parameter,
validates its conformity to the query's results, and normalizes the read
rows according to the provided schema.

This modification eliminates the need for the caller to validate the
returned records against its own schema. Consequently, the
`importUsersFromDatabase` method is corrected to avoid mistakenly
performing a second normalization, as it did before.

Additionally, the `Records` method handles the reading of identity and
timestamp, returning them in the `Record` type alongside the record
properties, similar to how the `(*App).Users` and `(*File).ReadFunc`
methods do.

For #380