apis/connectors: fix parsing of identity and timestamp columns in files
During the parsing of `identity` and `timestamp` columns, the code was
incorrectly retrieving values from the column with an index
corresponding to the properties' index, instead of using the correct
column index.

Additionally, normalization was being performed based on the type of the
provided schema rather than the schema of the respective columns.