all: always use up-to-date schemas instead of storing them
The idea behind this commit is to ensure that the user no longer has to
worry about having schemas within Chichi that are different from those
actually present in the data warehouse. Storing the schemas for `users`,
`users_identities`, and `events` within Chichi seems to offer no
advantages, only complications in usage and implementation.

In the future, we will implement a cache to ensure that the schemas do
not need to be read every time. However, this is an optimization that
the user should not be concerned about.

So, this commit:

* Ensures that the server no longer stores schemas inside the workspace
  but reads them each time they are needed, ensuring they are always
  up-to-date.
* Removes the concept of "reload schemas" since, for the user, Chichi
  always reflects the schemas of the data warehouse tables that are
  actually present in the database.
* Ensures that when connecting to a data warehouse, schemas are no
  longer read immediately. Instead, they are read only in the context
  where they are actually required.
* Removes the ability to reload the schema from the `admin/schema` admin
  page, as the `users` schema returned by the APIs is always the one
  just read from the data warehouse.
* Removes the `reload-schemas` command from `chichi-cli`.
* Adds references to the issue
  https://github.com/open2b/chichi/issues/392, where there is a
  discussion about whether to internally handle the schema in those
  points within the data warehouse without exposing it to Chichi.