apis: handle the events schema in actions as any other action schema
Currently, the input schemas of actions on events are treated in a
special manner, as Chichi directly utilizes the event schema encoded in
the software instead of the input schema.

This modification results in handling the patterns of actions on events
like any other action schema, simplifying the logical by eliminating a
special case. Additionally, this change may potentially pave the way for
allowing users to customize the event schema, particularly its traits,
in the future.

Furthermore, we can consider optimizing by not saving the event schema
to the database. However, this is an internal software behavior not
exposed through the APIs.