admin: refactor the app context
This commit performs a refactor of the app context, changing how the
app's initial state is loaded and how refreshes of outdated information
are handled.

From a development perspective, the code has been simplified and
standardized with practices and naming conventions used throughout the
rest of the application.

From a user standpoint, the only visible changes are:
- There is now a white screen with a centered spinner, which is
  displayed for a minimum of 300 milliseconds in cases where operations
  require a complete reload of the state (such as when removing a
  workspace or changing the currently selected workspace).
- Prior to this commit, clicking on a link that redirected to the same
  page resulted in a full document reload in the browser. Now, the
  application itself handles this situation by completely reloading the
  entire state, without requiring any operation on the browser.