admin: prevent alert dialog from showing when mapping is not modified
Currently, in the editing and creation screen of an action, if an
attempt is made to change the editing mode of the mapping (mappings,
Javascript, or Python), an alert dialog is always displayed.

The dialog informs the user that proceeding will result in the loss of
work done on the mapping in the current editing mode and then asks for
confirmation to continue.

This commit ensures that the dialog is not shown every time but only if
changes have actually been made to the mapping in the current editing
mode.