apis/transformers: fix tests of Float(32) values
Previously, the expected value for the "Float32" property was assigned
as a constant to a float64 variable, resulting in a stored value of
57.16038. When not assigned as a constant, the stored value became
57.16038131713867 instead of 57.16038.

Both values, when converted to float32, yield 57.16038. However, it is
more accurate to store 57.16038131713867, as that is the got value. The
test did not fail because the comparison occurs after the conversion to
float32.

In the event of a test error, the dumped values would have been
different.