apis: validate the data passed to `TransformPreview` against its schema
Previously, the data passed to `TransformPreview` was normalized upon
reading from an app, rather than being validated.

This commit introduces a new `apis/encoding` package with functions to
unmarshal and validate data serialized in JSON format, as well as
functions to marshal records into a serialized JSON format.

Additionally, it modifies the `TransformPreview` method to receive and
return data in the JSON serialized format instead of a `map[string]any`
value.

For #380