apis/mappings/mapexp: enable transformation of multiple expressions
Previously, the `mapexp` package only supported compiling and evaluating
single expressions. This change introduces a new `Transformer` type that
allows the compilation of multiple expressions together. The evaluation,
now referred to as 'transform,' returns a map[string]any value.

The new `Transformer.Transform` method returns an error if the
transformation of an expression fails.