apis/backoff: set attempts and cap with specific methods
The backoff already uses a context to control when to stop, so using
attempts should be an option. Consequently, this change moves
the attempts parameters from the `New` function to the specific
`SetLimits` method. Furthermore, It does the same for the cap and adds
a method to set the base.