apis/events: remove "active", "context"
This changes ensures that the event schema correspond to the events
returned by the data warehouse and not the events collected. As a
consequence, it removes the "groupId" and "traits" fields from the
context, and rearranges the session fields in a "session" object.

It also removes the "active" field of event because it is not handled as
expected.

Co-authored-by: Gianluca Mondini <mondini@open2b.com>