apis/datastore: return schema along with records
Previously, when retrieving records passing the invalid schema, the
returned records conformed to the table's schema, but the caller did not
have information about the schema. This change ensures that the `Select`
method of the data warehouse's driver and the `Users` and `Events`
methods of the store now return the schema along with the records.