apis: store the internal matching property in the action's input schema
Store the internal matching property of an action, if present, as a
string instead of a `types.Property`. So, the property with that name
must be present in the input schema of the action.

Thanks to this modification, the workaround code inside
`readUsersFromDataWarehouse` that may add the internal matching property
to the schema passed to the `store.Users` method can be removed, as this
property must already be present in that schema.