apis,connectors: use the default value if a property is not passed
Previously, if a value for a property was not provided for a database
export row, it would be assigned as nil. This could lead to an error in
query execution if the corresponding column was declared as "NOT NULL".

Now, instead of passing a value for the property to the `Upsert` method,
rows are passed as a map rather than a slice. The `Upsert` method
ensures that if a value for a property is not present, the default value
of the column, as defined in the table, is assigned.

As a result, it is now necessary for table columns to have a default
value.

Close #418